actually have shared tsx
